Move GSModel to sub-package
This will make it easier to organize the many classes.
