+++ src/test/java/org/opengis/cite/cdb10/cdbStructure/VerifyGSModelGeometryStructureTests.java
@@ -12,7 +12,7 @@ import static java.nio.file.StandardOpenOption.TRUNCATE_EXISTING;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.opengis.cite.cdb10.cdbStructure.GSModelGeometryStructureTests;
+import org.opengis.cite.cdb10.cdbStructure.GSModel.GSModelGeometryStructureTests;
 
 public class VerifyGSModelGeometryStructureTests extends StructureTestFixture<GSModelGeometryStructureTests> {
